# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Battery Research Workbench API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/experiments": {
      "get": {
        "operationId": "list_experiments_v2_api_v1_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "battery_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Battery Id"
            }
          },
          {
            "in": "query",
            "name": "is_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Experiments V2 Api V1 Experiments Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiments V2",
        "tags": [
          "intake",
          "experiments"
        ]
      },
      "post": {
        "operationId": "create_experiment_api_v1_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Experiment Api V1 Experiments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment",
        "tags": [
          "intake",
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{battery_id}/{experiment_id}": {
      "get": {
        "operationId": "get_experiment_api_v1_experiments__battery_id___experiment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Experiment Api V1 Experiments  Battery Id   Experiment Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Experiment",
        "tags": [
          "experiments"
        ]
      },
      "patch": {
        "operationId": "patch_experiment_api_v1_experiments__battery_id___experiment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Experiment Api V1 Experiments  Battery Id   Experiment Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Experiment",
        "tags": [
          "intake",
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{battery_id}/{experiment_id}/evidence": {
      "get": {
        "operationId": "evidence_api_v1_experiments__battery_id___experiment_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evidence Api V1 Experiments  Battery Id   Experiment Id  Evidence Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{battery_id}/{experiment_id}/limitations": {
      "get": {
        "operationId": "limitations_api_v1_experiments__battery_id___experiment_id__limitations_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Limitations Api V1 Experiments  Battery Id   Experiment Id  Limitations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Limitations",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{battery_id}/{experiment_id}/lineage": {
      "get": {
        "operationId": "lineage_api_v1_experiments__battery_id___experiment_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Lineage Api V1 Experiments  Battery Id   Experiment Id  Lineage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lineage",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{battery_id}/{experiment_id}/results": {
      "get": {
        "operationId": "results_api_v1_experiments__battery_id___experiment_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "result_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Results Api V1 Experiments  Battery Id   Experiment Id  Results Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Results",
        "tags": [
          "experiments"
        ]
      }
    },
    "/api/v1/experiments/{battery_id}/{experiment_id}/status": {
      "get": {
        "operationId": "get_status_api_v1_experiments__battery_id___experiment_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Status Api V1 Experiments  Battery Id   Experiment Id  Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Status",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{battery_id}/{experiment_id}/workspace-summary": {
      "get": {
        "operationId": "workspace_summary_api_v1_experiments__battery_id___experiment_id__workspace_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Workspace Summary Api V1 Experiments  Battery Id   Experiment Id  Workspace Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace Summary",
        "tags": [
          "experiments"
        ]
      }
    },
...
  },
...
}